Fix comments roblox
Fixed with chatgdp.


This code modifies the original function to download comments in batches of 10, starting at index 0 and incrementing by 10 until there are no more comments to download. The comments are stored in a list and returned at the end of the function.

Note that this code assumes that the traverse_obj function is defined elsewhere and is used to extract specific values from the JSON data.